% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/st_example.R
\name{st_example}
\alias{st_example}
\title{Spark geometry example.}
\usage{
st_example(sc, geom = "polygons")
}
\description{
Spark geometry example.
}
\details{
geometry can be  "polygons" or "points"
}
\examples{
\dontrun{
library(sparklyr)
sc <- spark_connect(master = "local")
register_gis(sc)
polygons_wkt <- st_example(sc, "polygons")
point_wkt <- st_example(sc, "points")
}
}
