% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/st_join.R
\name{st_join}
\alias{st_join}
\title{spatial join}
\usage{
st_join(x, y, join = NULL)
}
\arguments{
\item{x}{a spark spatial data frame}

\item{y}{a spark spatial data frame}
}
\value{
a spark spatial data frame, joined based on geometry
}
\description{
st_join: spatial join
}
\details{
alternative values for argument join are: ST_Contains, ST_Intersects, ST_Within, ST_Equals, ST_Crosses, ST_Touches, ST_Overlaps, ST_Distance
}
\examples{
\dontrun{

library(dplyr)
polygons <- read.table(system.file(package="geospark","examples/polygons.txt"), sep="|", col.names=c("area","geom"))
points <- read.table(system.file(package="geospark","examples/points.txt"), sep="|", col.names=c("city","state","geom"))
polygons_wkt <- copy_to(sc, polygons)
points_wkt <- copy_to(sc, points)
polygons_wkt <- mutate(polygons_wkt, y = st_geomfromwkt(geom))
points_wkt <- mutate(points_wkt, x = st_geomfromwkt(geom))
sc_res = st_join(polygons_wkt, points_wkt, join = sql("st_contains(y,x)"))
    group_by(area, state) \%>\%
    summarise(cnt = n()) 
}
}
