% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/st_example.R
\name{st_pit_example}
\alias{st_pit_example}
\title{Spark points example.}
\usage{
st_pit_example(sc)
}
\description{
Spark points example.
}
\examples{
\dontrun{
library(sparklyr)
sc <- spark_connect(master = "local")
register_gis(sc)
st_pit_example(sc)
}
}
