% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/st_example.R
\name{st_plg_example}
\alias{st_plg_example}
\title{Spark polygons example.}
\usage{
st_plg_example(sc)
}
\description{
Spark polygons example.
}
\examples{
\dontrun{
library(sparklyr)
sc <- spark_connect(master = "local")
register_gis(sc)
st_plg_example(sc)
}
}
