% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/register_gis.R
\name{register_gis}
\alias{register_gis}
\title{Enable GIS SQL}
\usage{
register_gis(sc)
}
\arguments{
\item{sc}{a spark connection

Used by 'sparklyr' to initilize GIS SQL.}
}
\value{
a GIS spark connection
}
\description{
Enable GIS SQL
}
\examples{
library(geospark)
library(sparklyr)

sc <- spark_connect(master = "spark://HOST:PORT")

# spark_connect() calls register_gis() automatically, as in:
register_gis(sc)

}
